requirements.txt
streamlit
supabase
pandas
psycopg2-binary
python-dotenv
plotly